Fix program crashing with DC-state
Problem was related to unclear status of District of Columbia as a US state, leading to confusion with the 'us' package devs.  Without the environment variable that I added, DC exists in package namespace but not as an actual state. This was apparently not the case when the base code was last developed.