Add docstring documentation to data.py
Add docstring documentation to data.py. Also noted that Hospital_US class might be used by code, added comment to it. Depending on what Emre says about the masters thesis using data, we might need to make an issue and ensure that the class works. Unclear whether the data is available or not.